+++ RELEASENOTES.md
@@ -84,6 +84,10 @@ Release notes
         ([#112](https://github.com/androidx/media/issues/112)).
     *   Don't start a foreground service with a pending intent for pause
         ([#167](https://github.com/androidx/media/issues/167)).
+    *   Manually hide the 'badge' associated with the notification created by
+        `DefaultNotificationProvider` on API 26 and API 27 (the badge is
+        automatically hidden on API 28+)
+        ([#131](https://github.com/androidx/media/issues/131)).
 *   RTSP:
     *   Add H263 fragmented packet handling
         ([#119](https://github.com/androidx/media/pull/119)).